update rot files
- add config file editor
- move telemetry out of minethd
