+++ version.h
@@ -1,4 +1,4 @@
 #pragma once
 
-#define XMR_STAK_NAME "xmr-stak-cpu"
-#define XMR_STAK_VERSION "1.3.0-1.5.0-dev"
+#define XMR_STAK_NAME "xmr-stak"
+#define XMR_STAK_VERSION "2.0.0-predev"
